ubik: Introduce new vote lock
Introduce a new lock to protect ubik data related to voting.
Specifically, it protects the following globals:
	ubik_lastYesTime
	lastYesHost
	lastYesClaim
	lastYesState
	lowestHost
	lowestTime
	syncHost
	syncTime
	ubik_dbVersion
	ubik_dbTid

Variables are grouped along with the lock in a new structure.

Also introduce a few helper functions to safely deal with ubik_dbVersion:
uvote_eq_dbVersion: Return true if the passed version is equal to the
current ubik_dbVersion
uvote_set_dbVersion: Set ubik_dbVersion to a specified value

Change-Id: I9bb248d0dfedc363181661ea723cac0af4928644
Reviewed-on: http://gerrit.openafs.org/4156
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>
Reviewed-by: Derrick Brashear <shadow@dementia.org>