Pull in 64BIT_ENV for FBSD
Support for some newer FreeBSD versions was added after we killed
AFS_64BIT_ENV, but the commit to kill it was not cherry-picked
to openafs-stable-1_6_x.  Cherry-picking the new version support
thus introduced a bug, as we still need to define AFS_64BIT_ENV
for these systems on this branch.

We attempted to fix this previously, but that fix was incomplete.
Apply the full fix now.

A direct commit to openafs-stable-1_6_x, as this change is not
relevant for master.

Change-Id: If83edd51552f4719ed80179f7d7423ab15bbb0cb
Reviewed-on: http://gerrit.openafs.org/3976
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>